+++ src/main/java/roomescape/dto/request/ReservationRequest.java
@@ -1,0 +1,8 @@
+package roomescape.dto.request;
+
+public record ReservationRequest(
+        String name,
+        String date,
+        Long timeId
+) {
+}
