+++ src/main/java/roomescape/dto/response/ReservationResponse.java
@@ -1,0 +1,19 @@
+package roomescape.dto.response;
+
+import roomescape.domain.Reservation;
+
+public record ReservationResponse(
+        Long id,
+        String name,
+        String date,
+        TimeResponse timeResponse
+) {
+    public static ReservationResponse from(Reservation reservation) {
+        return new ReservationResponse(
+                reservation.getId(),
+                reservation.getName(),
+                reservation.getDate(),
+                TimeResponse.from(reservation.getTime())
+        );
+    }
+}
