+++ src/main/java/roomescape/dto/response/TimeResponse.java
@@ -1,0 +1,12 @@
+package roomescape.dto.response;
+
+import roomescape.domain.ReservationTime;
+
+public record TimeResponse(
+        Long id,
+        String startAt
+) {
+    public static TimeResponse from(ReservationTime time) {
+        return new TimeResponse(time.getId(), time.getStartAt());
+    }
+}
